parser-python/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
openpyxl==3.1.2
celery==5.4.0
redis==5.0.4
openai==1.33.0
python-dotenv==1.0.1
# 향후 Gemini 및 Claude 연동용
google-generativeai==0.5.4
anthropic==0.26.1
# DB ORM 저장용
SQLAlchemy==2.0.30
PyMySQL==1.1.0
cryptography==42.0.5
